requirements.txt
Flask
sentence-transformers
faiss-cpu
google-generativeai
gunicorn